requirements.txt
datetime
python-dotenv
PyYaml
minimalmodbus
pyserial
